ath79: take over the space for user data on Linksys EA4500 v3
The last "syscfg" partition of the OEM firmware turns out to be a
UBIFS used to store user data, just as the "rootfs_data" of OpenWrt,
so it should be reasonable to absorb it into the "ubi" partition.

Factory installations via either OEM firmware or tftp are confirmed
working, but the UBI remaining inside "syscfg" partition could break
upgrade. Fortunately, installing kmod-mtd-rw and erasing "syscfg"
partition before upgrade is confirmed working.

Signed-off-by: Edward Chow <equu@openmail.cc>