tools/gnulib: unmangle fts header on macOS
The gnulib fts.h header is meant to not be overwritten
in any way by the host system's copy of fts.h
and is therefore given a unique name instead.

This is fine if the built libgnu library is directly linked
with the target library, but if we want to keep them isolated
we end up having the definitions being mangled anyway
when the next object to link against included the fts.h header.

On some macOS platforms, the use of __DARWIN_INODE64
is messing with the link name for fts functions, resulting in:

Undefined symbols for architecture x86_64:
  "_rpl_fts_close$INODE64", referenced from:
...

Create a local fts header for gnulib
that completely blocks the macOS host fts header.

Signed-off-by: Michael Pratt <mcpratt@pm.me>